test(collections): extract Collections REST API assert functions
